refactor suggestions-common.ts
https://coveord.atlassian.net/browse/KIT-5017
