+++ packages/atomic/src/components/common/suggestions/suggestions-common.spec.ts
@@ -1,165 +1,40 @@
-import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest';
-import {closest} from '../../../utils/dom-utils';
+import {describe, expect, it} from 'vitest';
 import {
   dispatchSearchBoxSuggestionsEvent,
   elementHasNoQuery,
   elementHasQuery,
   type SearchBoxSuggestionElement,
+  type SearchBoxSuggestions,
   type SearchBoxSuggestionsEvent,
 } from './suggestions-common';
 
-vi.mock('../../../utils/event-utils', {spy: true});
-vi.mock('../../../utils/dom-utils', {spy: true});
-
-describe('suggestions-common', () => {
-  const createTestElement = (query?: string): SearchBoxSuggestionElement => ({
-    key: 'test-key',
-    content: document.createElement('div'),
-    ...(query !== undefined && {query}),
-  });
-
-  describe('#elementHasNoQuery', () => {
-    it('should return true when element has undefined query', () => {
-      const element = createTestElement(undefined);
-      expect(elementHasNoQuery(element)).toBe(true);
+describe('utils', () => {
+  describe('re-exports', () => {
+    it('should export utility functions', () => {
+      expect(typeof elementHasNoQuery).toBe('function');
+      expect(typeof elementHasQuery).toBe('function');
     });
 
-    it('should return true when element has no query property', () => {
-      const element = createTestElement();
-      expect(elementHasNoQuery(element)).toBe(true);
+    it('should export event dispatching function', () => {
+      expect(typeof dispatchSearchBoxSuggestionsEvent).toBe('function');
     });
 
-    it('should return true when element has empty string query', () => {
-      const element = createTestElement('');
-      expect(elementHasNoQuery(element)).toBe(true);
-    });
+    it('should export types', () => {
+      const element: SearchBoxSuggestionElement = {
+        key: 'test',
+        content: document.createElement('div'),
+      };
 
-    it('should return false when element has non-empty query', () => {
-      const element = createTestElement('search query');
-      expect(elementHasNoQuery(element)).toBe(false);
-    });
+      const suggestions: SearchBoxSuggestions = {
+        position: 1,
+        renderItems: () => [element],
+      };
 
-    it('should return true when element has whitespace-only query', () => {
-      const element = createTestElement('   ');
-      expect(elementHasNoQuery(element)).toBe(true);
-    });
+      const event: SearchBoxSuggestionsEvent<unknown> = () => suggestions;
 
-    it('should return true when element has tab and newline characters', () => {
-      const element = createTestElement('\t\n  \r');
-      expect(elementHasNoQuery(element)).toBe(true);
-    });
-  });
-
-  describe('#elementHasQuery', () => {
-    it('should return false when element has undefined query', () => {
-      const element = createTestElement(undefined);
-      expect(elementHasQuery(element)).toBe(false);
-    });
-
-    it('should return false when element has no query property', () => {
-      const element = createTestElement();
-      expect(elementHasQuery(element)).toBe(false);
-    });
-
-    it('should return false when element has empty string query', () => {
-      const element = createTestElement('');
-      expect(elementHasQuery(element)).toBe(false);
-    });
-
-    it('should return true when element has non-empty query', () => {
-      const element = createTestElement('search query');
-      expect(elementHasQuery(element)).toBe(true);
-    });
-
-    it('should return false when element has whitespace-only query', () => {
-      const element = createTestElement('   ');
-      expect(elementHasQuery(element)).toBe(false);
-    });
-
-    it('should return false when element has tab and newline characters', () => {
-      const element = createTestElement('\t\n  \r');
-      expect(elementHasQuery(element)).toBe(false);
-    });
-
-    it('should return true when element has query with leading/trailing whitespace', () => {
-      const element = createTestElement('  search query  ');
-      expect(elementHasQuery(element)).toBe(true);
-    });
-
-    it('should handle numeric zero as valid query', () => {
-      const element = createTestElement('0');
-      expect(elementHasQuery(element)).toBe(true);
-    });
-  });
-
-  describe('#dispatchSearchBoxSuggestionsEvent', () => {
-    let mockElement: HTMLElement;
-    let mockEvent: SearchBoxSuggestionsEvent<unknown>;
-
-    beforeEach(() => {
-      document.body.innerHTML = '';
-      mockElement = document.createElement('div');
-      mockEvent = vi.fn();
-    });
-
-    afterEach(() => {
-      document.body.innerHTML = '';
-    });
-
-    it('should call closest with correct selector', () => {
-      vi.mocked(closest).mockReturnValue(
-        document.createElement('atomic-search-box')
-      );
-
-      dispatchSearchBoxSuggestionsEvent(mockEvent, mockElement);
-
-      expect(closest).toHaveBeenCalledWith(
-        mockElement,
-        'atomic-search-box, atomic-insight-search-box, atomic-commerce-search-box'
-      );
-    });
-
-    it('should work with any valid search box type', () => {
-      const searchBoxTypes = [
-        'atomic-search-box',
-        'atomic-insight-search-box',
-        'atomic-commerce-search-box',
-      ];
-
-      searchBoxTypes.forEach((type) => {
-        const mockParent = document.createElement(type);
-        vi.mocked(closest).mockReturnValue(mockParent);
-
-        expect(() => {
-          dispatchSearchBoxSuggestionsEvent(mockEvent, mockElement);
-        }).not.toThrow();
-      });
-    });
-
-    it('should throw error when element is not child of allowed search box elements', () => {
-      vi.mocked(closest).mockReturnValue(null);
-
-      expect(() => {
-        dispatchSearchBoxSuggestionsEvent(mockEvent, mockElement);
-      }).toThrow(
-        'The "div" component was not handled, as it is not a child of the following elements: atomic-search-box, atomic-insight-search-box, atomic-commerce-search-box'
-      );
-    });
-
-    it('should normalize nodeName to lowercase in error message', () => {
-      const upperCaseElement = document.createElement('div');
-      Object.defineProperty(upperCaseElement, 'nodeName', {
-        value: 'CUSTOM-ELEMENT',
-        writable: false,
-        configurable: true,
-      });
-      vi.mocked(closest).mockReturnValue(null);
-
-      expect(() => {
-        dispatchSearchBoxSuggestionsEvent(mockEvent, upperCaseElement);
-      }).toThrow(
-        'The "custom-element" component was not handled, as it is not a child of the following elements: atomic-search-box, atomic-insight-search-box, atomic-commerce-search-box'
-      );
+      expect(element).toBeDefined();
+      expect(suggestions).toBeDefined();
+      expect(event).toBeDefined();
     });
   });
 });
